Trial with a fake output that does the integration.
As we had it, the integration would try to get all the data points
that needed to be integrated in one go, which is guaranteed to run
out of memory for longer sample times.  Here, instead, I pass a fake
output to the raw read, which adds elements in the right place.