Delete rather than close ih
It was really not that logical to close all underlying filehandles.
One should just delete the reference, and if it is the last one held,
the file will be deleted and thus closed.

This way, one can use the sample input in multiple pipelines without
worrying about the file closing under one's feet.